PASSED MF!!!
Bejeezes this took waaaay to freaking long to figure out. 
I had it down until the "." that was being passed to my string.find() as an 'any character' which triggered one stupid freaking erroneous match. 
But I freaking got it!!
I spent so damn long trying to get the damn ptrn to fucking work but I fucking got it.